Refactor config module: split config.rs into focused modules
Split the monolithic src/config.rs (986 lines) into a well-organized
config/ module directory following the pattern from pool refactoring:

New structure:
- config/types.rs (135 lines)
  * RoutingMode enum with support queries
  * Config, ServerConfig structs
  * CacheConfig, HealthCheckConfig structs

- config/defaults.rs (50 lines)
  * All default value functions for serde deserialization
  * Centralized default constants

- config/validation.rs (112 lines)
  * Config::validate() implementation
  * validate_server(), validate_health_check(), validate_cache()
  * Keepalive interval warnings

- config/loading.rs (136 lines)
  * load_config() - TOML file loading
  * load_servers_from_env() - Environment variable support
  * create_default_config() - Test/example configurations

- config/mod.rs (20 lines)
  * Module coordination and re-exports
  * Clean public API

Total: 453 lines (down from 986, -54% reduction)

Benefits:
- Clear separation: types, defaults, validation, loading
- Each file has a single, focused responsibility
- Easier to navigate and maintain
- Better module organization
- Tests remain comprehensive (in integration tests)

Updated references:
- Renamed default_health_check_* functions to health_check_*
- Updated proxy.rs, pool/prewarming.rs, and test files

All tests passing.
