+++ src/pool/prewarming.rs
@@ -174,8 +174,8 @@ mod tests {
             tls_verify_cert: true,
             tls_cert_path: None,
             connection_keepalive_secs: 0,
-            health_check_max_per_cycle: crate::config::default_health_check_max_per_cycle(),
-            health_check_pool_timeout_ms: crate::config::default_health_check_pool_timeout_ms(),
+            health_check_max_per_cycle: crate::config::health_check_max_per_cycle(),
+            health_check_pool_timeout_ms: crate::config::health_check_pool_timeout_ms(),
         }];
 
         let providers = servers
