+++ src/proxy.rs
@@ -336,9 +336,7 @@ mod tests {
     use std::sync::Arc;
 
     fn create_test_config() -> Config {
-        use crate::config::{
-            default_health_check_max_per_cycle, default_health_check_pool_timeout_ms,
-        };
+        use crate::config::{health_check_max_per_cycle, health_check_pool_timeout_ms};
         Config {
             servers: vec![
                 ServerConfig {
@@ -352,8 +350,8 @@ mod tests {
                     tls_verify_cert: true,
                     tls_cert_path: None,
                     connection_keepalive_secs: 0,
-                    health_check_max_per_cycle: default_health_check_max_per_cycle(),
-                    health_check_pool_timeout_ms: default_health_check_pool_timeout_ms(),
+                    health_check_max_per_cycle: health_check_max_per_cycle(),
+                    health_check_pool_timeout_ms: health_check_pool_timeout_ms(),
                 },
                 ServerConfig {
                     host: "server2.example.com".to_string(),
@@ -366,8 +364,8 @@ mod tests {
                     tls_verify_cert: true,
                     tls_cert_path: None,
                     connection_keepalive_secs: 0,
-                    health_check_max_per_cycle: default_health_check_max_per_cycle(),
-                    health_check_pool_timeout_ms: default_health_check_pool_timeout_ms(),
+                    health_check_max_per_cycle: health_check_max_per_cycle(),
+                    health_check_pool_timeout_ms: health_check_pool_timeout_ms(),
                 },
                 ServerConfig {
                     host: "server3.example.com".to_string(),
@@ -380,8 +378,8 @@ mod tests {
                     tls_verify_cert: true,
                     tls_cert_path: None,
                     connection_keepalive_secs: 0,
-                    health_check_max_per_cycle: default_health_check_max_per_cycle(),
-                    health_check_pool_timeout_ms: default_health_check_pool_timeout_ms(),
+                    health_check_max_per_cycle: health_check_max_per_cycle(),
+                    health_check_pool_timeout_ms: health_check_pool_timeout_ms(),
                 },
             ],
             ..Default::default()
