+++ tests/config_helpers.rs
@@ -15,8 +15,8 @@ pub fn create_test_server_config(host: &str, port: u16, name: &str) -> ServerCon
         tls_verify_cert: true,
         tls_cert_path: None,
         connection_keepalive_secs: 0,
-        health_check_max_per_cycle: nntp_proxy::config::default_health_check_max_per_cycle(),
-        health_check_pool_timeout_ms: nntp_proxy::config::default_health_check_pool_timeout_ms(),
+        health_check_max_per_cycle: nntp_proxy::config::health_check_max_per_cycle(),
+        health_check_pool_timeout_ms: nntp_proxy::config::health_check_pool_timeout_ms(),
     }
 }
 
@@ -39,8 +39,8 @@ pub fn create_test_server_config_with_auth(
         tls_verify_cert: true,
         tls_cert_path: None,
         connection_keepalive_secs: 0,
-        health_check_max_per_cycle: nntp_proxy::config::default_health_check_max_per_cycle(),
-        health_check_pool_timeout_ms: nntp_proxy::config::default_health_check_pool_timeout_ms(),
+        health_check_max_per_cycle: nntp_proxy::config::health_check_max_per_cycle(),
+        health_check_pool_timeout_ms: nntp_proxy::config::health_check_pool_timeout_ms(),
     }
 }
 
@@ -63,8 +63,8 @@ pub fn create_test_server_config_with_tls(
         tls_verify_cert,
         tls_cert_path,
         connection_keepalive_secs: 0,
-        health_check_max_per_cycle: nntp_proxy::config::default_health_check_max_per_cycle(),
-        health_check_pool_timeout_ms: nntp_proxy::config::default_health_check_pool_timeout_ms(),
+        health_check_max_per_cycle: nntp_proxy::config::health_check_max_per_cycle(),
+        health_check_pool_timeout_ms: nntp_proxy::config::health_check_pool_timeout_ms(),
     }
 }
 
@@ -86,8 +86,8 @@ pub fn create_test_server_config_with_max_connections(
         tls_verify_cert: true,
         tls_cert_path: None,
         connection_keepalive_secs: 0,
-        health_check_max_per_cycle: nntp_proxy::config::default_health_check_max_per_cycle(),
-        health_check_pool_timeout_ms: nntp_proxy::config::default_health_check_pool_timeout_ms(),
+        health_check_max_per_cycle: nntp_proxy::config::health_check_max_per_cycle(),
+        health_check_pool_timeout_ms: nntp_proxy::config::health_check_pool_timeout_ms(),
     }
 }
